+++ docker-compose.yml
@@ -6,10 +6,10 @@ services:
     container_name: meowshop_db
     environment:
       - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=123
+      - POSTGRES_PASSWORD=postgres
       - POSTGRES_DB=meowshop
     ports:
-      - "5435:5432"
+      - "5432:5432"
     volumes:
       - postgres_data:/var/lib/postgresql/data
       - ./postgres/init:/docker-entrypoint-initdb.d
@@ -18,11 +18,6 @@ services:
       interval: 5s
       timeout: 5s
       retries: 5
-    restart: always
-
-  redis:
-    image: redis:7
-    container_name: meowshop_redis
     restart: always
 
   backend:
@@ -34,12 +29,9 @@ services:
       - "8000:8000"
     environment:
       - DATABASE_URL=postgresql://postgres:postgres@db:5432/meowshop
-      - REDIS_URL=redis://redis:6379
     depends_on:
       db:
         condition: service_healthy
-      redis:
-        condition: service_started
     volumes:
       - ./backend:/app
     restart: always
